Replace files from app to addon
