+++ addon/controllers/i-i-s-caseberry-logging-objects-application-log-e.js
@@ -1,0 +1,7 @@
+import EditFormController from 'ember-flexberry/controllers/edit-form';
+
+export default EditFormController.extend({
+    // Caption of this particular edit form.
+    title: 'Application log'
+
+});
