+++ addon/controllers/i-i-s-caseberry-logging-objects-application-log-l.js
@@ -1,0 +1,4 @@
+import ListFormController from 'ember-flexberry/controllers/list-form';
+
+var IISCaseberryLoggingObjectsApplicationLogLController = ListFormController;
+export default IISCaseberryLoggingObjectsApplicationLogLController;
