+++ addon/routes/i-i-s-caseberry-logging-objects-application-log-e.js
@@ -1,0 +1,6 @@
+import EditFormRoute from 'ember-flexberry/routes/edit-form';
+
+export default EditFormRoute.extend( {
+  modelProjection: 'ApplicationLogE',
+  modelName: 'i-i-s-caseberry-logging-objects-application-log'
+});
