+++ addon/routes/i-i-s-caseberry-logging-objects-application-log-l.js
@@ -1,0 +1,6 @@
+import ListFormRoute from 'ember-flexberry/routes/list-form';
+
+export default ListFormRoute.extend( {
+  modelProjection: 'ApplicationLogL',
+  modelName: 'i-i-s-caseberry-logging-objects-application-log'
+});
