--- app/controllers/i-i-s-caseberry-logging-objects-application-log-e/new.js
@@ -1,4 +1,0 @@
-import IISCaseberryLoggingObjectsApplicationLogEController from '../i-i-s-caseberry-logging-objects-application-log-e';
-
-var IISCaseberryLoggingObjectsApplicationLogENewController = IISCaseberryLoggingObjectsApplicationLogEController;
-export default IISCaseberryLoggingObjectsApplicationLogENewController;
