+++ app/routes/i-i-s-caseberry-logging-objects-application-log-e.js
@@ -1,6 +1,1 @@
-import EditFormRoute from 'ember-flexberry/routes/edit-form';
-
-export default EditFormRoute.extend( {
-  modelProjection: 'ApplicationLogE',
-  modelName: 'i-i-s-caseberry-logging-objects-application-log'
-}); 
+export { default } from 'ember-flexberry/routes/i-i-s-caseberry-logging-objects-application-log-e';
