--- app/routes/i-i-s-caseberry-logging-objects-application-log-e/new.js
@@ -1,6 +1,0 @@
-import EditFormNewRoute from 'ember-flexberry/routes/edit-form-new';
-
-export default EditFormNewRoute.extend( {
-  modelProjection: 'ApplicationLogE',
-  modelName: 'i-i-s-caseberry-logging-objects-application-log'
-});
