+++ app/routes/i-i-s-caseberry-logging-objects-application-log-l.js
@@ -1,6 +1,1 @@
-import ListFormRoute from 'ember-flexberry/routes/list-form';
-
-export default ListFormRoute.extend( {
-  modelProjection: 'ApplicationLogL',
-  modelName: 'i-i-s-caseberry-logging-objects-application-log'
-});
+export { default } from 'ember-flexberry/routes/i-i-s-caseberry-logging-objects-application-log-l';
